move ASSETS to constants/index.ts
