+++ src/constants/index.ts
@@ -1,3 +1,5 @@
+import type { Coin } from "../utils/types";
+
 export const enum TradeButtonStatus {
   Trade = "Trade",
   EnterAmount = "Enter an amount",
@@ -28,11 +30,6 @@ export const LIQUID_BTC = '6f0279e9ed041c3d710a9f57d0c02928416460c4b722ae3457a11
 export const LIQUID_USDT = 'ce091c998b83c78bb71a632313ba3760f1763d9cfcffae02258ffa9865a37bd2';
 export const LIQUID_CAD = '0e99c1a6da379d1f4151fb9df90449d40d0608f6cb33a5bcbfc8c265f42bab0a';
 
-// testnet
-const LIQUID_BTC_TESTNET = '144c654344aa716d6f3abcc1ca90e5641e4e2a7f633bc09fe3baf64585819a49';
-// regtest
-const LIQUID_BTC_REGTEST = '5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225';
-
 export const EXPLORER = 'https://blockstream.info/liquid/api'
 
 // used on FiatValue component
@@ -44,3 +41,32 @@ export const CoinGeckoId: Record<string,string> = {
   [Fiat.EUR]: "eur",
   [Fiat.USD]: "usd",
 };
+
+// ASSETS is an array of "featured assets" 
+// (i.e assets data harcoded in the app, it lets to save some network requests for very common assets)
+export const ASSETS: Coin[] = [
+  {
+    name: "Liquid bitcoin",
+    ticker: "L-BTC",
+    assetHash: LIQUID_BTC,
+    precision: 8,
+  },
+  {
+    ticker: 'USDT',
+    assetHash: LIQUID_USDT,
+    precision: 8,
+    name: 'Tether USD',
+  },
+  {
+    ticker: 'LCAD',
+    assetHash: LIQUID_CAD,
+    precision: 8,
+    name: 'Liquid CAD',
+  },
+  {
+    ticker: 'BTSE',
+    assetHash: 'b00b0ff0b11ebd47f7c6f57614c046dbbd204e84bf01178baf2be3713a206eb7',
+    precision: 8,
+    name: 'BTSE Token',
+  },
+];
